requirements.txt
neo4j==4.4.11
python-docx==0.8.11
streamlit==1.22.0
numpy==1.24.2
requests==2.28.2
python-dotenv==1.0.0